# Conflicts:
#	apps/api/src/main.rs
#	apps/api/src/routes/coach_log.rs
#	apps/web/src/routes/coach.tsx
#	apps/web/src/routes/nutrition.tsx
#	apps/web/src/routes/workouts.tsx
#	package-lock.json

+++ apps/api/src/models/daily_log.rs
@@ -8,10 +8,10 @@ pub struct DailyLog {
     pub id: Uuid,
     pub user_id: Uuid,
     pub date: NaiveDate,
-    pub water: i16,
-    pub sleep: i16,
+    pub water: i32,
+    pub sleep: f32,
     pub steps: i32,
-    pub protein_g: i16,
+    pub protein_g: f32,
     pub workout_done: bool,
     pub weight_kg: Option<f32>,
     pub created_at: Option<DateTime<Utc>>,
@@ -21,16 +21,10 @@ pub struct DailyLog {
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct UpdateDailyLog {
-    #[serde(alias = "water")]
-    pub water: Option<i16>,
-    #[serde(alias = "sleep")]
-    pub sleep: Option<i16>,
-    #[serde(alias = "steps")]
+    pub water: Option<i32>,
+    pub sleep: Option<f32>,
     pub steps: Option<i32>,
-    #[serde(alias = "protein_g")]
-    pub protein_g: Option<i16>,
-    #[serde(alias = "workout_done")]
+    pub protein_g: Option<f32>,
     pub workout_done: Option<bool>,
-    #[serde(alias = "weight_kg")]
     pub weight_kg: Option<f32>,
 }
